% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BVS.R
\name{BVS}
\alias{BVS}
\title{Bayesian Variable Selection}
\usage{
BVS(
  Y,
  NA_Y,
  X,
  XB,
  gammaCurr,
  BCurr,
  ZCurr,
  RCurr,
  invRCurr,
  d,
  responseType,
  pFixed,
  nCat,
  cutPoint,
  negBinomPar,
  hyperPar,
  samplerPar
)
}
\description{
Internal function that performs one Markov chain Monte Carlo iteration to sample from the posterior distribution of
all unknowns involved in the Bayesian Variable Selection step. For details, see \insertCite{Alexopoulos2021;textual}{MR2} and
\insertCite{Zuber2023;textual}{MR2}
}
\references{
\insertAllCited{}
}
