% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDR_BMM.R
\name{FDR_BMM}
\alias{FDR_BMM}
\title{False Discovery Rate based on Beta densities finite mixture model}
\usage{
FDR_BMM(PPI, w, FDR_nominal = 0.05, plotting = FALSE)
}
\arguments{
\item{PPI}{Ordered marginal posterior probability of inclusion (mPPI) or edge marginal posterior of inclusion (ePPI) return by
MBB_EM()}

\item{w}{Matrix of allocation probabilities}

\item{FDR_nominal}{Nominal level of False Discovery Rate to be controlled. Default values set at \code{0.05}}

\item{plotting}{Logical with no plots as default \code{FALSE}}
}
\value{
The value returned is a list object \code{list(cutoff, cutoff_idx, FDR_est, FNR_est)}
\itemize{
\item{\code{cutoff}}{ Posterior probability of inclusion cutoff for a given nominal FDR level}
\item{\code{cutoff_idx}}{ Index of the ordered posterior probability of inclusion that corresponds to the estimated cutoff }
\item{\code{FDR_est}}{ Estimated False Discovery Rate }
\item{\code{FNR_est}}{ Estimated False Negative Rate} }
}
\description{
Post-processing of MR2 Markov chain Monte Carlo output. Estimation of the posterior probability of inclusion cutoff
for a given nominal False Discovery Rate level
}
\details{
For details regarding the False Discovery Rate control based on the Beta densities finite mixture model, see in
\insertCite{Zuber2023;textual}{MR2}
}
\examples{
# Example 1: Analysis of one replication of simulated Scenario II-Confounding with q = 5 
# responses, p = 15 exposures and nIV = 100 genetic variants used as IVs. The number of 
# expected exposures directly associated with each response is set at 2 and its variance at 2, 
# with a priori range of direct causal association ranging between 0 and 8

Sim_Confounding <- Scenario_Confounding(nRep = 1, q = 5, p = 15, nIV = 100, seed = 280610971)
betaHat_Y <- Sim_Confounding$betaHat_Y
betaHat_X <- Sim_Confounding$betaHat_X

MR2_output <- MR2(betaHat_Y, betaHat_X, EVgamma = c(2, 2), 
                  niter = 7500, burnin = 2500, thin = 5, monitor = 1000, seed = 280610971)

PostProc_output <- PostProc(MR2_output, betaHat_Y, betaHat_X)

BMM_EM_output <- BMM_EM(PostProc_output$gammaPost)
FDR_BMM_output <- FDR_BMM(BMM_EM_output$PPI, BMM_EM_output$w)
PostProc_output$thetaPost * (PostProc_output$gammaPost > FDR_BMM_output$cutoff) * 1


# Example 2: Analysis of one replication of simulated Scenario IV-Directed pleiotropy with q = 
# 5 responses, p = 15 exposures and nIV = 100 genetic variants used as IVs and the effect of 
# the shared pleiotropic pathway on the responses set as \code{2}. The number of expected 
# exposures directly associated with each response is set at 1 and its variance at 2, 
# with a priori range of direct causal associations ranging between 0 and 7. A non-decomposable 
# graph for the inverse of the residual correlation between responses is selected

Sim_Pleiotropy <- Scenario_Pleiotropy(nRep = 1, q = 5, p = 15, nIV = 100, undirectedA = FALSE, 
                                      seed = 280610971)
betaHat__Y <- Sim_Pleiotropy$betaHat_Y
betaHat__X <- Sim_Pleiotropy$betaHat_X

MR2_output <- MR2(betaHat_Y, betaHat_X, EVgamma = c(1, 3),
                  niter = 15000, burnin = 5000, thin = 10, monitor = 1000, 
                  nonDecomp = TRUE, seed = 280610971)

PostProc_output <- PostProc(MR2_output, betaHat_Y, betaHat_X)

BMM_EM_output <- BMM_EM(PostProc_output$gammaPost)
FDR_BMM_output <- FDR_BMM(BMM_EM_output$PPI, BMM_EM_output$w)
PostProc_output$thetaPost * (PostProc_output$gammaPost > FDR_BMM_output$cutoff) * 1
}
\references{
\insertAllCited{}
}
