% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_BDgraph.R
\name{mod_BDgraph}
\alias{mod_BDgraph}
\title{Birth and Death sampler of non-decomposable graphs}
\usage{
mod_BDgraph(
  data,
  g.start,
  K.start,
  samplerPar,
  n = NULL,
  not.cont = NULL,
  df.prior = 2,
  g.prior = 0.5,
  threshold = 1e-08
)
}
\description{
Internal function that performs one Markov chain Monte Carlo iteration to sample from the posterior distribution of
a non-decomposable graph. The R code of this function is a modified version of the BDgraph R-package
(\insertCite{Mohammadi2019;textual}{MR2}). For details, see also \insertCite{Mohammadi2015;textual}{MR2} and
\insertCite{Alexopoulos2021;textual}{MR2}
}
\references{
\insertAllCited{}
}
