% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MR2.R
\name{MR2}
\alias{MR2}
\title{MR2: Multi-response Mendelian randomization}
\usage{
MR2(
  betaHat_Y,
  betaHat_X,
  EVgamma = c(2, 2),
  tau = 1,
  niter,
  burnin,
  thin,
  monitor,
  fullCov = FALSE,
  nonDecomp = FALSE,
  alpha = 0.05,
  probAddDel = 0.9,
  probAdd = 0.5,
  probMix = 0.25,
  geomMean = EVgamma[1],
  graphParNiter = max(16, dim(Y)[2] * dim(Y)[2]),
  seed = 31122021
)
}
\arguments{
\item{betaHat_Y}{(number of IVs) times (number of responses) matrix of summary-level responses}

\item{betaHat_X}{(number of IVs) times (number of exposures) matrix of summary-level exposures}

\item{EVgamma}{A priori number of expected exposures directly associated with each response and its variance. For details, see
\insertCite{Kohn2001;textual}{MR2}, \insertCite{Alexopoulos2021;textual}{MR2} and \insertCite{Zuber2023;textual}{MR2}. Default
value is \code{EVgamma = c(2, 2)} which implies a priori range of important exposures associated with each response between 0 and
6. If a single value is provided between (0, 1), it corresponds to the prior probability of exposure-outcome causal association.
It is compulsory when a single exposure is employed}

\item{tau}{Prior variance of the (direct causal effect) theta prior with the default value set at \code{1}. Do not change this
value unless prior knowledge of the prior variance is available}

\item{niter}{Number of Markov chain Monte Carlo iterations (including burn-in)}

\item{burnin}{Number of Markov chain Monte Carlo iterations to be discarded as burn-in}

\item{thin}{Store the Markov chain Monte Carlo output at every thin-th iteration}

\item{monitor}{Display the monitored-th iteration}

\item{fullCov}{Logical parameter to select full (\code{TRUE}) or sparse (\code{FALSE} default) inverse correlation matrix
Markov chain Monte Carlo sampler}

\item{nonDecomp}{Logical, \code{TRUE} if a non-decomposable graphical model for the inverse correlation matrix is selected}

\item{alpha}{Level of significance (\code{0.05} default) to select important exposures for each response by using one at-a-time
response, univariable MR. Used in the construction of the proposal distribution of \code{gamma}}

\item{probAddDel}{Probability (\code{0.9} default) of adding/deleting one exposure to be directly associated in a model with a
response during Markov chain Monte Carlo exploration}

\item{probAdd}{Probability (\code{0.9} default) of adding one exposure to be directly associated in a model with a response during
Markov chain Monte Monte Carlo exploration}

\item{probMix}{Probability (\code{0.25} default) of selecting a geometric proposal distribution that samples the index of the
exposure being added/deleted. For details, see \insertCite{Alexopoulos2021;textual}{MR2}}

\item{geomMean}{Mean of the geometric proposal distribution that samples the index of the exposure being added/deleted. Default
value is the number of expected exposures directly associated with each response and specified in \code{EVgamma}. For details,
see \insertCite{Alexopoulos2021;textual}{MR2}}

\item{graphParNiter}{Number of Markov chain Monte Carlo internal iterations to sample the graphical model between responses.
Default value is the max between 16 and the square of the number of responses}

\item{seed}{Seed used to initialise the Markov chain Monte Carlo algorithm}
}
\value{
The value returned is a list object \code{list(theta, G, R, D, postMean, hyperPar, samplerPar, opt)}
\itemize{
\item{\code{theta}}{ Matrix of the (thinned) samples drawn from the posterior distribution of the direct causal effects }
\item{\code{G}}{ 3D array of the (thinned) samples drawn from the posterior distribution of the graphs }
\item{\code{R}}{ 3D array of the (thinned) samples drawn from the posterior distribution of the correlation matrix between the
responses }
\item{\code{D}}{ Matrix of the (thinned) samples drawn from the posterior distribution of the standard deviations of each
responses }
\item{\code{postMean}}{ List of the posterior means of \code{list(gamma, theta, G, R, D)} }
\item{\code{timeMR2}}{ Time in minutes employed by \code{MR2} algorithm to analyse the data }
\item{\code{hyperPar}}{ List of the hyper-parameters \code{list(tau)} and the parameters of the beta-binomial distribution
derived from \code{EVgamma} }
\item{\code{samplerPar}}{ List of parameters used in the  Markov chain Monte Carlo algorithm
\code{list(alpha, probAddDel, probAdd, probMix, geomMean, graphParNiter)} }
\item{\code{opt}}{ List of options used \code{list(std, seed)} } }
}
\description{
Markov chain Monte Carlo implementation of Bayesian Multi-response Mendelian randomization model for two-sample
summary-level data
}
\details{
For details regarding the model and the algorithm, see \insertCite{Zuber2023;textual}{MR2} and
\insertCite{Alexopoulos2021;textual}{MR2}, respectively
}
\examples{
# Example 1: Analysis of one replication of simulated Scenario II-Confounding with q = 5 
# responses, p = 15 exposures and nIV = 100 genetic variants used as IVs. The number of expected 
# exposures directly associated with each response is set at 2 and its variance at 2, with a 
# priori range of direct causal associations ranging between 0 and 8

Sim_Confounding <- Scenario_Confounding(nRep = 1, q = 5, p = 15, nIV = 100, seed = 280610971)
betaHat_Y <- Sim_Confounding$betaHat_Y
betaHat_X <- Sim_Confounding$betaHat_X

MR2_output <- MR2(betaHat_Y, betaHat_X, EVgamma = c(2, 2), 
                  niter = 7500, burnin = 2500, thin = 5, monitor = 1000, seed = 280610971)
head(MR2_output$postMean$theta)


# Example 2: Analysis of one replication of simulated Scenario IV-Directed pleiotropy with q = 
# 5 responses, p = 15 exposures and nIV = 100 genetic variants used as IVs and the effect of 
# the shared pleiotropic pathway on the responses set as \code{2}. The number of expected 
# exposures directly associated with each response is set at 1 and its variance at 2, 
# with a priori range of direct causal associations ranging between 0 and 7. A non-decomposable 
# graph for the inverse of the residual correlation between responses is selected

Sim_Pleiotropy <- Scenario_Pleiotropy(nRep = 1, q = 5, p = 15, nIV = 100, undirectedA = FALSE, 
                                      seed = 280610971)
betaHat_Y <- Sim_Pleiotropy$betaHat_Y
betaHat_X <- Sim_Pleiotropy$betaHat_X

MR2_output <- MR2(betaHat_Y, betaHat_X, EVgamma = c(1, 3),
                  niter = 15000, burnin = 5000, thin = 10, monitor = 1000, 
                  nonDecomp = TRUE, seed = 280610971)
head(MR2_output$postMean$theta)


}
\references{
\insertAllCited{}
}
