% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PostProc.R
\name{PostProc}
\alias{PostProc}
\title{Post-processing of MR2 output}
\usage{
PostProc(output, betaHat_Y, betaHat_X, alpha = 0.05, digits = 3)
}
\arguments{
\item{output}{Object returned by MR2}

\item{betaHat_Y}{(number of IVs) times (number of responses) matrix of summary-level responses}

\item{betaHat_X}{(number of IVs) times (number of exposures) matrix of summary-level exposures}

\item{alpha}{Area of the tails of a symmetric credible interval (\code{0.05} default) centred around the estimated direct
causal effects and the estimated covariance and (partial) residual correlation matrices}

\item{digits}{Integer indicating the number of decimal places}
}
\value{
The value returned is a list object \code{list(gammaPost, thetaPost, thetaPost_CI, DPost, GPost, SigmaPost, SigmaPost_CI, 
RPost, RPost_CI, Rm1Post, Rm1Post_CI, decomposability_freq, CPO, scaleCPO, freq_ouliers, PLML, DIC, opt)}
\itemize{
\item{\code{gammaPost}}{ Marginal posterior probability of inclusion (mPPI) }
\item{\code{thetaPost}}{ Posterior mean of the direct causal effects }
\item{\code{thetaPost_CI}}{  of the direct causal effects }
\item{\code{DPost}}{ Posterior mean of the standard deviations of each response }
\item{\code{GPost}}{ Edge posterior probability of inclusion (ePPI) }
\item{\code{SigmaPost}}{ Posterior mean of the variance-covariance matrix between the responses after conditioning on the direct
causal effects }
\item{\code{SigmaPost_CI}}{ (1 - \eqn{\alpha}) credible interval of the variance-covariance matrix between the responses after
conditioning on the direct causal effects }
\item{\code{RPost}}{ Posterior mean of the correlation matrix between the responses after conditioning on the direct causal
effects }
\item{\code{RPost_CI}}{ (1 - \eqn{\alpha}) credible interval of the correlation matrix between the responses after conditioning on the
direct causal effects }
\item{\code{Rm1Post}}{ Posterior mean of the partial correlation matrix between the responses after conditioning on the direct
causal effects }
\item{\code{Rm1Post_CI}}{ (1 - \eqn{\alpha}) credible interval of the partial correlation matrix between the responses after
conditioning on the direct causal effects }
\item{\code{decomposability_freq}}{ Frequency of visited decomposability models during the Markov chain Monte Carlo. If
\code{nonDecomp = TRUE}, frequency can be less than 1 }
\item{\code{CPO}}{ Conditional predictive ordinate for each instrumental variable, see \insertCite{Ntzoufras2008;textual}{MR2} }
\item{\code{scaleCPO}}{ Scaled conditional predictive ordinate for each instrumental variable, see
\insertCite{Congdon2005;textual}{MR2} }
\item{\code{PLML}}{ Pseudo log-marginal likelihood}
\item{\code{scaleCPO}}{ Scaled conditional predictive ordinate for each instrumental variable, see
\insertCite{Congdon2005;textual}{MR2} }
\item{\code{freq_ouliers}}{ Frequency outliers or high-leverage and influential instrumental variables defined as the frequency
of scale CPO < 0.01, see \insertCite{Congdon2005;textual}{MR2} }
\item{\code{stability_PLML}}{ Variance Pseudo log-marginal likelihood }
\item{\code{DIC}}{ Deviation Information Criteria, see \insertCite{Spiegelhalter2002;textual}{MR2} }
\item{\code{timePostProc}}{ Time in minutes employed by \code{PostProc} to perform the analyse }
\item{\code{opt}}{ List of options used \code{list(\eqn{\alpha}, digits)} } }
}
\description{
Post-processing of MR2 Markov chain Monte Carlo output
}
\details{
For details regarding the post-processing of the Markov chain Monte Carlo output, see details in
\insertCite{Zuber2023;textual}{MR2}
}
\examples{
# Example 1: Analysis of one replication of simulated Scenario II-Confounding with q = 5 
# responses, p = 15 exposures and nIV = 100 genetic variants used as IVs. The number of 
# expected exposures directly associated with each response is set at 2 and its variance at 2, 
# with a priori range of direct causal association ranging between 0 and 8

Sim_Confounding <- Scenario_Confounding(nRep = 1, q = 5, p = 15, nIV = 100, seed = 280610971)
betaHat_Y <- Sim_Confounding$betaHat_Y
betaHat_X <- Sim_Confounding$betaHat_X

MR2_output <- MR2(betaHat_Y, betaHat_X, EVgamma = c(2, 2), 
                  niter = 7500, burnin = 2500, thin = 5, monitor = 1000, seed = 280610971)
PostProc_output <- PostProc(MR2_output, betaHat_Y, betaHat_X)


# Example 2: Analysis of one replication of simulated Scenario IV-Directed pleiotropy with q = 
# 5 responses, p = 15 exposures and nIV = 100 genetic variants used as IVs and the effect of 
# the shared pleiotropic pathway on the responses set as \code{2}. The number of expected 
# exposures directly associated with each response is set at 1 and its variance at 3, 
# with a priori range of direct causal associations ranging between 0 and 7. A non-decomposable 
# graph for the inverse of the residual correlation between responses is selected

Sim_Pleiotropy <- Scenario_Pleiotropy(nRep = 1, q = 5, p = 15, nIV = 100, undirectedA = FALSE, 
                                      seed = 280610971)
betaHat_Y <- Sim_Pleiotropy$betaHat_Y
betaHat_X <- Sim_Pleiotropy$betaHat_X

MR2_output <- MR2(betaHat_Y, betaHat_X, EVgamma = c(1, 3),
                  niter = 15000, burnin = 5000, thin = 10, monitor = 1000, 
                  nonDecomp = TRUE, seed = 280610971)

PostProc_output <- PostProc(MR2_output, betaHat_Y, betaHat_X)
}
\references{
\insertAllCited{}
}
