% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sample_G.R
\name{Sample_G}
\alias{Sample_G}
\title{Sampler of decomposable graphs}
\usage{
Sample_G(W, GCurr, samplerPar, delta = 2)
}
\description{
Internal functions that perform one Markov chain Monte Carlo iteration to sample from the posterior distribution of a
decomposable graph. The R code of this function is a modified version of the Matlab code in \insertCite{Talhouk2012;textual}{MR2}.
For details, see \insertCite{Alexopoulos2021;textual}{MR2}
}
\references{
\insertAllCited{}
}
