% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sample_gamma.R
\name{Sample_gamma}
\alias{Sample_gamma}
\title{Sampler of the binary latent variable}
\usage{
Sample_gamma(gammaCurr, p, mlogpval, samplerPar)
}
\description{
Internal function for the proposal distribution of the binary latent vector for each response based on
\insertCite{Guan2011;textual}{MR2}. For details, see \insertCite{Alexopoulos2021;textual}{MR2}
}
\references{
\insertAllCited{}
}
