% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scenario_Mediation.R
\name{Scenario_Mediation}
\alias{Scenario_Mediation}
\title{Simulated scenario: Mediation}
\usage{
Scenario_Mediation(
  nRep = 2,
  q = 5,
  p = 10,
  nIV = 100,
  nSubject = 50000,
  MAF = 0.05,
  thetaRange = c(-2, 2),
  thetaSparsness = 0.3,
  heritY = 0.25,
  rhoY = 0.6,
  betaHatRange_X = c(-2, 2),
  heritX = 0.1,
  rhoX = 0.6,
  undirectedA = TRUE,
  thetaA = 1,
  thetaM = 1,
  thetaUY = 1,
  thetaUX = 2,
  seed = 31122021
)
}
\arguments{
\item{nRep}{Number of simulated replicates}

\item{q}{Number of simulated summary-level responses}

\item{p}{Number of simulated summary-level exposures}

\item{nIV}{Number of genetic variants used as instrumental variables}

\item{nSubject}{Number of individuals in each sample of individual-level data}

\item{MAF}{Major allele frequency used in the simulation of the genotypes in each sample individual-level data. Default value set
at \code{0.05}}

\item{thetaRange}{Sought range of the simulated direct causal effects. Default value set at \code{(-2,2)}}

\item{thetaSparsness}{Proportion of simulated causal associations between exposures and responses. Default value set at \code{0.30}}

\item{heritY}{Heriditability of the responses. Default value set at \code{0.25}}

\item{rhoY}{Correlation between the responses not due to any unmeasured pleiotropic pathway. Default value set at \code{0.60}}

\item{betaHatRange_X}{Sought range of the simulated genetic effects on the exposures. Default value set at \code{(-2,2)}}

\item{heritX}{Heriditability of the exposures. Default value set at \code{0.10}}

\item{rhoX}{Correlation between the exposures. Default value set at \code{0.60}}

\item{undirectedA}{Logical parameter for simulating an undirected (\code{TRUE}, default) or directed pleiotropic pathway}

\item{thetaA}{Effect of the shared pleiotropic pathway on the responses. Default value set at \code{1}}

\item{thetaM}{Effect of one response on another one, both selected at random. Default value set at \code{1}}

\item{thetaUY}{Impact of the confounder on the responses. Default value set at \code{1}}

\item{thetaUX}{Impact of the confounder on the exposures. Default value set at \code{2}}

\item{seed}{Seed used to initialise the simulation}
}
\value{
The value returned is a list object \code{list(betaHat_Y, betaHat_X, theta, par)}
\itemize{
\item{\code{betaHat_Y}}{ 3D array (IVs times responses times replicates) of the simulated (summary-level) regression coefficients
of the  genetic effects on each response }
\item{\code{betaHat_X}}{ 3D array (IVs times exposures times replicates) of the simulated (summary-level) regression coefficients
of the genetic effects on each exposure }
\item{\code{theta}}{ Matrix ((exposures times replicates) times replicates) of the simulated direct causal effects }
\item{\code{mediation}}{ Matrix (two times replicates) times replicates) of the selected responses to be the mediated response
and the mediator }
\item{\code{par}}{ List of all parameters used in the simulation \code{list(nRep, q, p, nIV, nSubject, MAF, thetaRange, 
        thetaSparsness, heritY, rhoY, betaHatRange_X, heritX, rhoX, undirectedA, thetaA, thetaM, thetaUY, thetaUX, seed)} } }
}
\description{
Simulator to generate multiple responses and multiple exposures two-sample summary-level data with confounding and
pleiotropic effects as well as the effect of one response on another one
}
\details{
For details regarding the simulated scenario, see details in \insertCite{Zuber2023;textual}{MR2}
}
\examples{
# Example 1: Simulation of one replication of simulated Scenario Undirected pleiotropy with 
# Mediation and with q = 5 responses, p = 15 exposures and nIV = 100 genetic variants used as 
# IVs

Sim_Mediation <- Scenario_Mediation(nRep = 1, q = 5, p = 15, nIV = 100, seed = 280610971)
Sim_Mediation$mediation
head(Sim_Mediation$theta)
Sim_Mediation$par


# Example 2: Simulation of one replication of simulated Scenario Directed pleiotropy with 
# Mediation and with q = 5 responses, p = 15 exposures and nIV = 100 genetic variants used as 
# IVs and the effect of the shared pleiotropic pathway on the responses set as \code{2}

Sim_Mediation <- Scenario_Mediation(nRep = 1, q = 5, p = 15, nIV = 100, undirectedA = FALSE, 
                                    thetaA = 2, seed = 280610971)
Sim_Mediation$mediation
head(Sim_Mediation$theta)
Sim_Mediation$par
}
\references{
\insertAllCited{}
}
