% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim_HIW.R
\name{Sim_HIW}
\alias{Sim_HIW}
\title{Sampler of Hyper-Inverse Wishart}
\usage{
Sim_HIW(G, S, n)
}
\description{
Internal functions to simulate from the Hyper Inverse-Wishart distribution given a decomposable graph. The R code of
this function is a modified version of the Matlab code of \insertCite{Carvalho2007;textual}{MR2} and
\insertCite{Talhouk2012;textual}{MR2}. For details, see details, see \insertCite{Alexopoulos2021;textual}{MR2}
}
\references{
\insertAllCited{}
}
